# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AddableParticipantItem": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "name"
        ],
        "title": "AddableParticipantItem",
        "type": "object"
      },
      "AddableParticipantsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AddableParticipantItem"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "AddableParticipantsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JoinRequestBody": {
        "properties": {
          "trust_creator": {
            "default": false,
            "title": "Trust Creator",
            "type": "boolean"
          }
        },
        "required": [
          "trust_creator"
        ],
        "title": "JoinRequestBody",
        "type": "object"
      },
      "JoinStatusItem": {
        "properties": {
          "can_join": {
            "default": false,
            "title": "Can Join",
            "type": "boolean"
          },
          "can_leave": {
            "default": false,
            "title": "Can Leave",
            "type": "boolean"
          },
          "can_manage": {
            "default": false,
            "title": "Can Manage",
            "type": "boolean"
          },
          "can_record": {
            "default": false,
            "title": "Can Record",
            "type": "boolean"
          },
          "has_matches": {
            "default": false,
            "title": "Has Matches",
            "type": "boolean"
          },
          "is_participant": {
            "title": "Is Participant",
            "type": "boolean"
          }
        },
        "required": [
          "is_participant",
          "can_join",
          "can_leave",
          "can_manage",
          "can_record",
          "has_matches"
        ],
        "title": "JoinStatusItem",
        "type": "object"
      },
      "JoinStatusResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JoinStatusItem"
          }
        },
        "required": [
          "data"
        ],
        "title": "JoinStatusResponse",
        "type": "object"
      },
...
      "ParticipantAddBody": {
        "properties": {
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "ParticipantAddBody",
        "type": "object"
      },
...
      "ScorerAddBody": {
        "properties": {
          "user_email": {
            "title": "User Email",
            "type": "string"
          }
        },
        "required": [
          "user_email"
        ],
        "title": "ScorerAddBody",
        "type": "object"
      },
      "ScorerItem": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "name"
        ],
        "title": "ScorerItem",
        "type": "object"
      },
      "ScorersResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ScorerItem"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "ScorersResponse",
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "SuccessResponse",
        "type": "object"
      },
...
      "TournamentsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Tournament"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "TournamentsResponse",
        "type": "object"
      },
      "TrustedManagerAddBody": {
        "properties": {
          "manager_email": {
            "title": "Manager Email",
            "type": "string"
          }
        },
        "required": [
          "manager_email"
        ],
        "title": "TrustedManagerAddBody",
        "type": "object"
      },
      "TrustedManagerItem": {
        "properties": {
          "manager_id": {
            "title": "Manager Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "manager_id",
          "name"
        ],
        "title": "TrustedManagerItem",
        "type": "object"
      },
      "TrustedManagersResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TrustedManagerItem"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "TrustedManagersResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\n### Description\nThis API allows you to do everything the frontend of [Bracket](https://github.com/evroon/bracket)\nallows you to do (the frontend uses this API as well).\n\nFore more information, see the [documentation](https://docs.bracketapp.nl).\n\n### Table of Contents\n*(links only work for [ReDoc](https://api.bracketapp.nl/redoc), not for Swagger UI)*\n\n- [Auth](#tag/Auth)\n\n- [Clubs](#tag/Clubs)\n\n- [Courts](#tag/Courts)\n\n- [Favorites](#tag/Favorites)\n\n- [Internals](#tag/Internals)\n\n- [Matches](#tag/Matches)\n\n- [Participants](#tag/Participants)\n\n- [Players](#tag/Players)\n\n- [Rankings](#tag/Rankings)\n\n- [Ratings](#tag/Ratings)\n\n- [Rounds](#tag/Rounds)\n\n- [Showcase](#tag/Showcase)\n\n- [Stage Items](#tag/Stage-Items)\n\n- [Stage Item Inputs](#tag/Stage-Item-Inputs)\n\n- [Stages](#tag/Stages)\n\n- [Teams](#tag/Teams)\n\n- [Tournaments](#tag/Tournaments)\n\n- [Users](#tag/Users)\n\n### Links\nGitHub: <https://github.com/evroon/bracket>\n\nDocs: <https://docs.bracketapp.nl>\n\nDemo: <https://www.bracketapp.nl/demo>\n\nAPI docs (Redoc): <https://api.bracketapp.nl/redoc>\n\nAPI docs (Swagger UI): <https://api.bracketapp.nl/docs>\n",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "summary": "API for Bracket, an open source tournament system.",
    "title": "Bracket API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/me/joinable-tournaments": {
      "get": {
        "description": "Public individual tournaments the current user could join right now \u2014\nthe \u53ef\u4ee5\u53c2\u8d5b\u7684\u6bd4\u8d5b tab on the home page.",
        "operationId": "list_joinable_tournaments_me_joinable_tournaments_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Joinable Tournaments",
        "tags": [
          "Participants"
        ]
      }
    },
...
    "/me/trusted-managers": {
      "get": {
        "operationId": "list_trusted_managers_me_trusted_managers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustedManagersResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Trusted Managers",
        "tags": [
          "Participants"
        ]
      },
      "post": {
        "operationId": "add_my_trusted_manager_me_trusted_managers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrustedManagerAddBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add My Trusted Manager",
        "tags": [
          "Participants"
        ]
      }
    },
    "/me/trusted-managers/{manager_id}": {
      "delete": {
        "operationId": "remove_my_trusted_manager_me_trusted_managers__manager_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "manager_id",
            "required": true,
            "schema": {
              "title": "Manager Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Remove My Trusted Manager",
        "tags": [
          "Participants"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/addable-participants": {
      "get": {
        "operationId": "list_addable_participants_tournaments__tournament_id__addable_participants_get",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddableParticipantsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Addable Participants",
        "tags": [
          "Participants"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/join-requests": {
      "post": {
        "description": "A logged-in user joins an individual tournament directly \u2014 no creator\napproval, no rejection. Allowed only before play starts (no match generated).",
        "operationId": "request_to_join_tournaments__tournament_id__join_requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Request To Join",
        "tags": [
          "Participants"
        ]
      }
    },
    "/tournaments/{tournament_id}/leave": {
      "post": {
        "description": "A participant withdraws themselves \u2014 removing their single-person team and\nbinding. Allowed only before play starts (no match generated) and never after\nsettlement.",
        "operationId": "leave_tournament_tournaments__tournament_id__leave_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Leave Tournament",
        "tags": [
          "Participants"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/my-join-status": {
      "get": {
        "description": "The current user's relationship to this tournament: whether they already\nparticipate (a bound team), can manage it, and can record scores. Used to pick\nthe right \u53c2\u8d5b / \u9000\u51fa\u8d5b\u4e8b affordance.",
        "operationId": "my_join_status_tournaments__tournament_id__my_join_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "My Join Status",
        "tags": [
          "Participants"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/participants": {
      "post": {
        "operationId": "add_participant_tournaments__tournament_id__participants_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipantAddBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Participant",
        "tags": [
          "Participants"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/scorers": {
      "get": {
        "operationId": "list_scorers_tournaments__tournament_id__scorers_get",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Scorers",
        "tags": [
          "Participants"
        ]
      },
      "post": {
        "operationId": "add_tournament_scorer_tournaments__tournament_id__scorers_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScorerAddBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Tournament Scorer",
        "tags": [
          "Participants"
        ]
      }
    },
    "/tournaments/{tournament_id}/scorers/{user_id}": {
      "delete": {
        "operationId": "remove_tournament_scorer_tournaments__tournament_id__scorers__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Remove Tournament Scorer",
        "tags": [
          "Participants"
        ]
      }
    },
...
  }
}